+++ crosslinking_bot/utils.py
@@ -4,7 +4,7 @@ import urltools
 from urllib import parse
 
 
-def same_url(url1, url2):
+def same_url(raw_url1, raw_url2):
     """Check if 2 URLs refer to the same primary resource
 
     `urltools.compare()` fails if the 2 URLs have different fragments.
@@ -19,21 +19,32 @@ def same_url(url1, url2):
     Returns:
         bool: Whether the URLs are the same
     """
-    exception_path = '/blog'
     arxiv_exception = 'arxiv.org'
-    plos_exception = 'journals.plos.org'
+    query_field_exception = 'id'
 
-    _, netloc1, path1, params1, query1, fragment1 = parse.urlparse(url1)
-    _, netloc2, path2, params2, query2, fragment2 = parse.urlparse(url2)
-    # If the path is simply 'blog', judge from the fragment
-    if netloc1 == netloc2 and path1 == path2 == exception_path:
-        return fragment1 == fragment2
+    url1 = parse_url(raw_url1)
+    url2 = parse_url(raw_url2)
+
+    if url1['netloc'] == url2['netloc'] and url1['path'] == url2['path']:
+        # If one of the query field is `id`, parse the value
+        if query_field_exception in url1['query'] and query_field_exception in url2['query']:
+            return (parse.unquote(url1['query'][query_field_exception][0]) == 
+                    parse.unquote(url2['query'][query_field_exception][0]))
+        else:
+            return urltools.compare(url1['netloc'] + url1['path'], url2['netloc'] + url2['path'])
     # If it's on arxiv, do some acrobatics
-    elif netloc1 == netloc2 == arxiv_exception:
+    elif url1['netloc'] == url2['netloc'] == arxiv_exception:
         regex = '([^/a-z]+\.[^/a-z.]+)'
-        return re.findall(regex, path1) == re.findall(regex, path2)
-    # If it's on PLOS, compare the query
-    elif netloc1 == netloc2 == plos_exception:
-        return parse.unquote(query1) ==  parse.unquote(query2)
+        return re.findall(regex, url1['path']) == re.findall(regex, url2['path'])
     else:
-        return urltools.compare(netloc1 + path1, netloc2 + path2)
+        return urltools.compare(url1['netloc'] + url1['path'], url2['netloc'] + url2['path'])
+
+def parse_url(url):
+    """Parse URL into a dictionary"""
+    url_dict = {}
+
+    parsed_result = parse.urlparse(url)
+    for key, value in zip(parsed_result._fields, parsed_result):
+        url_dict[key] = value if key != 'query' else parse.parse_qs(value)
+
+    return url_dict